refactor(data): extraer el schema DDL a db/schema.cjs — fase c (05/T03)
createBaseSchema(db) (PRAGMAs + todas las tablas/índices/FTS/triggers,
IF NOT EXISTS, byte-idéntico al original verificado por diff) sale a
electron/core/db/schema.cjs. database.cjs baja a 657 líneas y queda como
fachada (getDB/initDatabase/getQueries/integridad/reparación/cascadas),
orquestando createBaseSchema → populateFTS → runMigrations →
createDefaultProject. Cierra 05/T03 (a: queries, b: migraciones, c:
schema). Sin cambios de comportamiento.
